feat: 1. I have added checking right of user for products. 2. The add_product has been the new checking "Group not found" . 3. The update_product has been the new checking "Group not found". 4. I need made the testing of API.

+++ app/controllers/product_controller.py
@@ -3,7 +3,10 @@ from flask_jwt_extended import jwt_required
 from sqlalchemy import select
 from ..models.group import Group
 from ..models.product import Product
+from ..models.user import User
+from ..models.rights_user import RightsUser
 from ..extensions import db
+from flask_jwt_extended import get_jwt_identity
 
 product_controller = Blueprint('product', __name__)
 
@@ -12,10 +15,17 @@ product_controller = Blueprint('product', __name__)
 @jwt_required()
 def get_products(group_id=None):
     try:
-        # need to add checking right of user
+        identity = get_jwt_identity()
+                
+        query = select(User).filter_by(username=identity)
+                        
+        user = db.session.execute(query).scalar_one_or_none()
+                
+        if not user:
+            return jsonify({"error": "Unauthorized"}), 401
 
         if group_id is None:
-            query = select(Product)
+            query = select(Product).join(RightsUser).filter(RightsUser.user_id == user.id)
             all_products = db.session.execute(query).scalars().all()
 
             return jsonify([{'id': p.id, 'name': p.name, 'group_id': p.group_id} for p in all_products])
@@ -26,10 +36,18 @@ def get_products(group_id=None):
         if not group:
             return jsonify({'error': 'Group not found'}), 404
 
-        query = select(Product).filter_by(group_id=group_id)
+        query = select(Group).join(RightsUser).filter(RightsUser.user_id == user.id, RightsUser.group_id == group_id)
+        group = db.session.execute(query).scalar_one_or_none()
+
+        if not group:
+            return jsonify({"error": "Forbidden"}), 403
+
+        # query = select(Product).filter_by(group_id=group_id)
+        query = select(Product).join(RightsUser).filter(RightsUser.user_id == user.id, RightsUser.group_id == group_id)
         products = db.session.execute(query).scalars().all()
 
-        query = select(Group).filter_by(parent_id=group_id)
+        # query = select(Group).filter_by(parent_id=group_id)
+        query = select(Group).join(RightsUser).filter(RightsUser.user_id == user.id, Group.parent_id == group_id)
         sub_groups = db.session.execute(query).scalars().all()
 
         data = {
@@ -63,13 +81,32 @@ def add_product():
     try:
         data = request.get_json()
 
-        # need to add checking right of user
+        identity = get_jwt_identity()
+                        
+        query = select(User).filter_by(username=identity)
+
+        user = db.session.execute(query).scalar_one_or_none()
+                        
+        if not user:
+            return jsonify({"error": "Unauthorized"}), 401
 
         if not data or 'name' not in data or 'group_id' not in data:
             return jsonify({"error": "Bad Request"}), 400
         
         if data['group_id'] is None:
             return jsonify({"error": "Bad Request"}), 400
+
+        parent = db.session.get(Group, data['group_id'])
+
+        if not parent:
+            return jsonify({"error": "Group not found"}), 404
+
+        query = select(RightsUser).filter_by(user_id=user.id, group_id=data['group_id'])
+
+        rights = db.session.execute(query).scalar_one_or_none()
+        
+        if not rights:
+            return jsonify({"error": "Forbidden"}), 403
 
         product = Product(name = data['name'], group_id = data['group_id'])
 
@@ -92,21 +129,38 @@ def update_product(id):
         if not product:
             return jsonify({"error": "Not found"}), 404
 
-        # need to add checking right of user
+        identity = get_jwt_identity()
+                                
+        query = select(User).filter_by(username=identity)
+        
+        user = db.session.execute(query).scalar_one_or_none()
+                                
+        if not user:
+            return jsonify({"error": "Unauthorized"}), 401
 
         data = request.get_json()
 
         if not data:
             return jsonify({"error": "Bad Request"}), 400
         
-        if 'name' in data:
-            product.name = data['name']
-        
-        if 'group_id' in data:
-            product.group_id = data['group_id']
+        if 'group_id' in data and 'name' in data:
+            if data['group_id'] is None:
+                return jsonify({"error": "Bad Request"}), 400
 
-        if data['group_id'] is None:
-            return jsonify({"error": "Bad Request"}), 400
+            parent = db.session.get(Group, data['group_id'])
+            
+            if not parent:
+                return jsonify({"error": "Group not found"}), 404
+
+            query = select(RightsUser).filter_by(user_id=user.id, group_id=data['group_id'])
+            
+            rights = db.session.execute(query).scalar_one_or_none()
+                    
+            if not rights:
+                return jsonify({"error": "Forbidden"}), 403
+
+            product.name = data['name']
+            product.group_id = data['group_id']
         
         db.session.commit()
 
@@ -122,10 +176,24 @@ def delete_product(id):
     try:
         product = db.session.get(Product, id)
 
-        # need to add checking right of user
+        identity = get_jwt_identity()
+                                        
+        query = select(User).filter_by(username=identity)
+                
+        user = db.session.execute(query).scalar_one_or_none()
+                                        
+        if not user:
+            return jsonify({"error": "Unauthorized"}), 401
 
         if not product:
             return jsonify({"error": "Not found"}), 404
+
+        query = select(RightsUser).filter_by(user_id=user.id, group_id=product.group_id)
+                    
+        rights = db.session.execute(query).scalar_one_or_none()
+                            
+        if not rights:
+            return jsonify({"error": "Forbidden"}), 403
         
         db.session.delete(product)
